refactor(audit): split middleware module
Mechanical split of middlewares/audit.rs into middlewares/audit/{mod,client_ip,extensions}.rs:
- extensions.rs: request-extension marker types (KmipOperationName, KmipObjectUid,
  KmipAlgorithm, BatchItemAuditContext, KmipBatchOperations), re-exported unchanged
  from mod.rs so external call sites (middlewares/mod.rs, routes/kmip/audit.rs) need
  no changes.
- client_ip.rs: extract_operation / extract_client_ip and their tests.
- mod.rs: AuditMiddleware/AuditService (Transform/Service impl), unchanged behavior.

No backend changes. Ran the log-reference.md generator (--non-interactive) to update
the 2 call-site paths that moved; left 3 unrelated pre-existing [REMOVED]-flagged
entries (src/routes/kmip/handlers.rs) untouched -- out of scope for this commit.
